完成大部分

+++ src/main/java/com/gs/controller/systemManage/ModuleController.java
@@ -4,6 +4,7 @@ import ch.qos.logback.classic.Logger;
 import com.gs.bean.Module;
 import com.gs.bean.Permission;
 import com.gs.common.bean.ControllerResult;
+import com.gs.common.util.RoleUtil;
 import com.gs.common.util.SessionUtil;
 import com.gs.common.util.UUIDUtil;
 import com.gs.controller.SystemManageController;
@@ -41,9 +42,11 @@ public class ModuleController {
     private ModuleService moduleService;
 
 
+    // 保留
     @ResponseBody
     @RequestMapping(value = "queryAll")
     public List<Module> queryAll(HttpSession session){
+        // TODO 所有人可以看
         if(SessionUtil.isLogin(session)) {
             return moduleService.queryAll("Y");
         } else {
@@ -52,39 +55,59 @@ public class ModuleController {
         }
     }
 
+    // baoliu
     @ResponseBody
     @RequestMapping(value = "insert")
     public Map insert(Module module, HttpSession session){
+        // TODO 超级管理员可以
         if(SessionUtil.isLogin(session)) {
-            Map map = new HashMap();
-            module.setModuleId(UUIDUtil.uuid());
-            int insertCount = moduleService.insert(module);
-            if(checkInsertOneSuccess(insertCount)) {
-                map.put("controllResult", ControllerResult.getSuccessResult("添加成功"));
-                map.put("module", module);
+            String roles = "系统超级管理员,系统普通管理员";
+            if(RoleUtil.checkRoles(roles)) {
+                Map map = new HashMap();
+                module.setModuleId(UUIDUtil.uuid());
+                int insertCount = moduleService.insert(module);
+                if(checkInsertOneSuccess(insertCount)) {
+                    map.put("controllResult", ControllerResult.getSuccessResult("添加成功"));
+                    map.put("module", module);
+                } else {
+                    map.put("controllResult", ControllerResult.getFailResult("添加失败"));
+                    map.put("module", null);
+                }
+                return map ;
             } else {
-                map.put("controllResult", ControllerResult.getFailResult("添加失败"));
-                map.put("module", null);
+                logger.info("此用户无拥有添加模块的角色");
+                Map map = new HashMap();
+                map.put("controllResult",ControllerResult.getNotRoleResult("权限不足"));
+                return map ;
             }
-            return map ;
         } else {
             logger.info("请先登录");
             return null;
         }
     }
 
+    // baoliu
     @ResponseBody
     @RequestMapping(value = "edit")
     public Map edit(Module module, HttpSession session) {
+        // TODO 超级管理员可以
         if(SessionUtil.isLogin(session)) {
-            int insertCount = moduleService.update(module);
-            Map map = new HashMap();
-            if (checkInsertOneSuccess(insertCount)) {
-                map.put("controllResult", ControllerResult.getSuccessResult("修改成功"));
+            String roles = "系统超级管理员,系统普通管理员";
+            if(RoleUtil.checkRoles(roles)) {
+                int insertCount = moduleService.update(module);
+                Map map = new HashMap();
+                if (checkInsertOneSuccess(insertCount)) {
+                    map.put("controllResult", ControllerResult.getSuccessResult("修改成功"));
+                } else {
+                    map.put("controllResult", ControllerResult.getFailResult("修改失败"));
+                }
+                return map;
             } else {
-                map.put("controllResult", ControllerResult.getFailResult("修改失败"));
+                logger.info("此用户无拥有修改模块的角色");
+                Map map = new HashMap();
+                map.put("controllResult",ControllerResult.getNotRoleResult("权限不足"));
+                return map;
             }
-            return map;
         } else {
             logger.info("请先登录");
             return null;
@@ -92,18 +115,26 @@ public class ModuleController {
     }
 
 
+    // baoliu
     @ResponseBody
     @RequestMapping(value = "delete")
     public ControllerResult delete(@RequestParam("moduleId") String moduleId, HttpSession session) {
+        // TODO 超级管理员可以
         if(SessionUtil.isLogin(session)) {
-            Module module = new Module();
-            module.setModuleId(moduleId);
-            int deleteCount = moduleService.delete(module);
-            moduleService.updPermissions(moduleId, null);
-            if(checkInsertOneSuccess(deleteCount)) {
-                return ControllerResult.getSuccessResult("删除成功");
+            String roles = "系统超级管理员,系统普通管理员";
+            if(RoleUtil.checkRoles(roles)) {
+                Module module = new Module();
+                module.setModuleId(moduleId);
+                int deleteCount = moduleService.delete(module);
+                moduleService.updPermissions(moduleId, null);
+                if(checkInsertOneSuccess(deleteCount)) {
+                    return ControllerResult.getSuccessResult("删除成功");
+                }
+                return ControllerResult.getFailResult("删除失败");
+            } else {
+                logger.info("此用户无拥有修改模块的角色");
+                return ControllerResult.getNotRoleResult("权限不足");
             }
-            return ControllerResult.getFailResult("删除失败");
         } else {
             logger.info("请先登录");
             return null;
@@ -111,15 +142,23 @@ public class ModuleController {
     }
 
 
+    // baoliu
     @ResponseBody
     @RequestMapping(value = "updPer")
     public ControllerResult updPer(@RequestParam("moduleId") String moduleId,@RequestParam("permissionId") String permissionId, HttpSession session) {
+        // TODO 超级管理员可以
         if(SessionUtil.isLogin(session)) {
-            int updPerCount = moduleService.updPermission(permissionId,moduleId);
-            if(checkInsertOneSuccess(updPerCount)) {
-                return ControllerResult.getSuccessResult("修改成功");
+            String roles = "系统超级管理员,系统普通管理员";
+            if(RoleUtil.checkRoles(roles)) {
+                int updPerCount = moduleService.updPermission(permissionId,moduleId);
+                if(checkInsertOneSuccess(updPerCount)) {
+                    return ControllerResult.getSuccessResult("修改成功");
+                }
+                return ControllerResult.getFailResult("修改失败");
+            } else {
+                logger.info("此用户无拥有修改模块的权限的角色");
+                return ControllerResult.getNotRoleResult("权限不足");
             }
-            return ControllerResult.getFailResult("修改失败");
         } else {
             logger.info("请先登录");
             return null;
